Move from G++ to Clang++ compilation

+++ src/engine/deal.cpp
@@ -9,9 +9,10 @@ void Deal::addLoan(std::string groupId, Loan &loan)
 
 unsigned int Deal::numberOfLoans() const
 {
-    return std::accumulate(
-        collateral.begin(),
-        collateral.end(),
-        0,
-        [](int count, std::pair<std::string, CollateralGroup> group) { return count + group.second.loans.size(); });
+    return static_cast<unsigned int>(
+        std::accumulate(
+            collateral.begin(),
+            collateral.end(),
+            0,
+            [](unsigned int count, std::pair<std::string, CollateralGroup> group) { return count + group.second.loans.size(); }));
 }
